# Fix Email Processing Logic and Performance Issues