--- a/tests/test_data_generator.py
+++ b/tests/test_data_generator.py
@@ -4,6 +4,7 @@ from src.gmail_client import Email
 
 # Do not modify this file
 
+
 class TestDataGenerator:
     def __init__(self):
         self.names = [