--- a/tests/test_email_processor.py
+++ b/tests/test_email_processor.py
@@ -86,7 +86,7 @@ class TestEmailProcessor:
         assert "filtered_emails" in result
         assert "responses_sent" in result
         assert result["total_emails"] == 10
-        
+
 
     def test_email_filtering_accuracy(self):
         valid_emails = []