--- a/tests/test_email_processor.py
+++ b/tests/test_email_processor.py
@@ -87,7 +87,6 @@ class TestEmailProcessor:
         assert "responses_sent" in result
         assert result["total_emails"] == 10
 
-
     def test_email_filtering_accuracy(self):
         valid_emails = []
         invalid_emails = []
@@ -176,7 +175,6 @@ class TestEmailProcessor:
         assert result["responses_sent"] == result["filtered_emails"]
         assert (end_time - start_time) < self.performance_threshold
         assert len(mock_client.sent_emails) == result["responses_sent"]
-
 
     def test_performance_with_1000_emails(self):
         test_emails = self.data_generator.generate_test_emails(1000)